feat: Update weather data sources and models for consistency and migration support

+++ Projet_8_Construisez_et_testez_une_infrastructure_de_données/greenandcoop/models/2_silver/slv_infoclimat.sql
@@ -9,25 +9,25 @@ select
     dh_utc::timestamp without time zone     as observed_at,
     'infoclimat'                            as source_system,
 
-    -- Mesures communes (ordre identique dans les 3 silver)
-    temperature                             as temperature_c,
-    pression                                as pressure_hpa,
-    humidite                                as humidity_pct,
-    point_de_rosee                          as dew_point_c,
-    round((vent_moyen * 3.6)::numeric, 1)   as wind_speed_kmh,
-    round((vent_rafales * 3.6)::numeric, 1) as wind_gust_kmh,
-    vent_direction                          as wind_direction_deg,
+    -- Mesures communes (castées en numeric pour compatibilité UNION)
+    temperature::numeric                    as temperature_c,
+    pression::numeric                       as pressure_hpa,
+    humidite::numeric                       as humidity_pct,
+    point_de_rosee::numeric                 as dew_point_c,
+    round((vent_moyen::numeric * 3.6)::numeric, 1)   as wind_speed_kmh,
+    round((vent_rafales::numeric * 3.6)::numeric, 1)  as wind_gust_kmh,
+    vent_direction::numeric                 as wind_direction_deg,
     null::text                              as wind_direction_text,
-    pluie_1h                                as precip_rate_mmh,
+    pluie_1h::numeric                       as precip_rate_mmh,
     null::numeric                           as precip_accum_mm,
-    pluie_3h                                as precip_accum_3h_mm,
+    pluie_3h::numeric                       as precip_accum_3h_mm,
     null::numeric                           as uv_index,
     null::numeric                           as solar_radiation_wm2,
 
     -- Mesures spécifiques infoclimat
-    visibilite                              as visibility_m,
-    nebulosite                              as cloud_cover_oktas,
-    neige_au_sol                            as snow_ground_cm,
-    temps_omm                               as weather_code_omm
+    visibilite::numeric                     as visibility_m,
+    nebulosite::numeric                     as cloud_cover_oktas,
+    neige_au_sol::numeric                   as snow_ground_cm,
+    temps_omm::numeric                      as weather_code_omm
 
 from source
